:recycle: Move data models out of ShareService
