+++ lib/pages/home/modals/save_post/create_post.dart
@@ -20,7 +20,9 @@ import 'package:Okuna/services/localization.dart';
 import 'package:Okuna/services/media/media.dart';
 import 'package:Okuna/services/media/models/media_file.dart';
 import 'package:Okuna/services/navigation_service.dart';
-import 'package:Okuna/services/share.dart';
+import 'package:Okuna/services/share/models/share.dart';
+import 'package:Okuna/services/share/models/subscription.dart';
+import 'package:Okuna/services/share/share.dart';
 import 'package:Okuna/services/toast.dart';
 import 'package:Okuna/services/user.dart';
 import 'package:Okuna/services/validation.dart';
