+++ lib/services/share/models/share.dart
@@ -1,0 +1,9 @@
+import 'dart:io';
+
+class Share {
+  final String text;
+  final File image;
+  final File video;
+
+  const Share({this.text, this.image, this.video});
+}
