+++ lib/services/share/models/subscription.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/cupertino.dart';
+
+class ShareSubscription {
+  final VoidCallback _cancel;
+
+  ShareSubscription(this._cancel);
+
+  void cancel() {
+    _cancel();
+  }
+}
